+++ submissions/week-2/day-3/banjo_o_s/garage_employees/src/employees/constants/constants.rs
@@ -1,0 +1,20 @@
+#[derive(Debug, PartialEq)]
+pub enum OtherEmployeeRole {
+    SocialMedia,
+    Technician,
+    Kitchen,
+}
+
+#[derive(Debug)]
+pub enum EmployeeStatus {
+    Terminated,
+    Engaged,
+}
+
+#[derive(Debug, PartialEq)]
+pub enum EmployeeRole {
+    Media,
+    IT,
+    Manager,
+    OtherEmployeeRole(u8),
+}
