+++ submissions/week-2/day-3/banjo_o_s/garage_employees/src/main.rs
@@ -1,5 +1,13 @@
 pub mod employees {
-    pub mod garage;
+    pub mod models {
+        pub mod models;
+    }
+    pub mod services {
+        pub mod services;
+    }
+    pub mod constants {
+        pub mod constants;
+    }
 }
 
 pub mod test {
